Now BlurAlgorithm can have float radius

+++ library/src/main/java/eightbitlab/com/blurview/BlurAlgorithm.java
@@ -8,7 +8,7 @@ public interface BlurAlgorithm {
      * @param blurRadius blur radius
      * @return blurred bitmap
      */
-    Bitmap blur(Bitmap bitmap, int blurRadius);
+    Bitmap blur(Bitmap bitmap, float blurRadius);
 
     /**
      * Frees allocated resources
